requirements.txt
streamlit>=1.28.0
yfinance>=0.2.32
pandas>=2.1.0
numpy>=1.26.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.3.0
ta>=0.10.2
vaderSentiment>=3.3.2
plotly>=5.18.0
pytz>=2023.3
flask>=2.3.0
flask-cors>=4.0.0
